Ejecutable/requirements.txt
apache-beam==2.32.0
flask==2.1.4
folium==0.12.1
google-cloud-bigquery==2.31.0
google-cloud-pubsub==3.3.0
pandas==1.3.3
pandas-gbq==0.15.0
plotly==5.3.1
requests==2.26.0
streamlit==1.1.0
streamlit-folium==0.5.0
